change order list to dva
